Switch PAC to stm32-rs/stm32wb crate.
This is likely to be more maintained going forward.
Only changes are due to upstream renaming fields in the SVD to be more
standard across families, and to more closely match the RefManuals.

Depends on either https://github.com/stm32-rs/stm32-rs/pull/624
or https://github.com/stm32-rs/stm32-rs/pull/473
and then a new upstream pac release, so _this_ commit should change to
depend on those, but this shows what's needed and how, for preview.

Signed-off-by: Karl Palsson <karlp@etactica.com>